Move logic from effect to payment connector
