+++ projects/core/src/cart/connectors/payment/cart-payment.adapter.ts
@@ -2,19 +2,15 @@ import { Observable } from 'rxjs';
 import { PaymentDetails } from '../../../occ/occ-models/occ.models';
 
 export abstract class CartPaymentAdapter {
-  abstract getProviderSubInfo(userId: string, cartId: string): Observable<any>;
-  abstract createSubWithProvider(
-    postUrl: string,
-    parameters: any
-  ): Observable<any>;
-  abstract createDetails(
+  abstract create(
     userId: string,
     cartId: string,
-    parameters: any
+    paymentDetails: PaymentDetails
   ): Observable<PaymentDetails>;
+
   abstract setDetails(
     userId: string,
     cartId: string,
-    paymentDetailsId: any
+    paymentDetailsId: string
   ): Observable<any>;
 }
