+++ projects/core/src/cart/connectors/payment/cart-payment.connector.ts
@@ -9,29 +9,18 @@ import { CartPaymentAdapter } from './cart-payment.adapter';
 export class CartPaymentConnector {
   constructor(private adapter: CartPaymentAdapter) {}
 
-  public getProviderSubInfo(userId: string, cartId: string): Observable<any> {
-    return this.adapter.getProviderSubInfo(userId, cartId);
-  }
-
-  public createSubWithProvider(
-    postUrl: string,
-    parameters: any
-  ): Observable<any> {
-    return this.adapter.createSubWithProvider(postUrl, parameters);
-  }
-
-  public createDetails(
+  public create(
     userId: string,
     cartId: string,
-    parameters: any
+    paymentDetails: PaymentDetails
   ): Observable<PaymentDetails> {
-    return this.adapter.createDetails(userId, cartId, parameters);
+    return this.adapter.create(userId, cartId, paymentDetails);
   }
 
   public setDetails(
     userId: string,
     cartId: string,
-    paymentDetailsId: any
+    paymentDetailsId: string
   ): Observable<any> {
     return this.adapter.setDetails(userId, cartId, paymentDetailsId);
   }
