+++ projects/core/src/cart/connectors/payment/index.ts
@@ -1,2 +1,3 @@
 export * from './cart-payment.connector';
 export * from './cart-payment.adapter';
+export * from './converters';
