+++ projects/core/src/checkout/store/effects/checkout.effect.ts
@@ -6,7 +6,7 @@ import * as fromActions from '../actions/index';
 import { AddMessage, GlobalMessageType } from '../../../global-message/index';
 import { ProductImageNormalizer } from '../../../product/index';
 import { OccOrderService } from '../../../user/index';
-import { OrderEntry, PaymentDetails } from '../../../occ/occ-models/index';
+import { OrderEntry } from '../../../occ/occ-models/index';
 import * as fromUserActions from '../../../user/store/actions/index';
 import { CartDeliveryConnector } from '../../../cart/connectors/delivery/cart-delivery.connector';
 import { CartPaymentConnector } from '../../../cart/connectors/payment/cart-payment.connector';
@@ -107,65 +107,17 @@ export class CheckoutEffects {
     mergeMap(payload => {
       // get information for creating a subscription directly with payment provider
       return this.cartPaymentConnector
-        .getProviderSubInfo(payload.userId, payload.cartId)
+        .create(payload.userId, payload.cartId, payload.paymentDetails)
         .pipe(
-          map(data => {
-            const labelsMap = this.convertToMap(data.mappingLabels.entry);
-            return {
-              url: data.postUrl,
-              parameters: this.getParamsForPaymentProvider(
-                payload.paymentDetails,
-                data.parameters.entry,
-                labelsMap
-              ),
-              mappingLabels: labelsMap,
-            };
+          mergeMap(details => {
+            return [
+              new fromUserActions.LoadUserPaymentMethods(payload.userId),
+              new fromActions.CreatePaymentDetailsSuccess(details),
+            ];
           }),
-          mergeMap(sub => {
-            // create a subscription directly with payment provider
-            return this.cartPaymentConnector
-              .createSubWithProvider(sub.url, sub.parameters)
-              .pipe(
-                map(response => this.extractPaymentDetailsFromHtml(response)),
-                mergeMap(fromPaymentProvider => {
-                  if (!fromPaymentProvider['hasError']) {
-                    // consume response from payment provider and creates payment details
-
-                    return this.cartPaymentConnector
-                      .createDetails(
-                        payload.userId,
-                        payload.cartId,
-                        this.getPaymentSopResponseParams(
-                          payload.paymentDetails,
-                          fromPaymentProvider,
-                          sub.mappingLabels
-                        )
-                      )
-                      .pipe(
-                        mergeMap(details => {
-                          return [
-                            new fromUserActions.LoadUserPaymentMethods(
-                              payload.userId
-                            ),
-                            new fromActions.CreatePaymentDetailsSuccess(
-                              details
-                            ),
-                          ];
-                        }),
-                        catchError(error =>
-                          of(new fromActions.CreatePaymentDetailsFail(error))
-                        )
-                      );
-                  } else {
-                    return of(
-                      new fromActions.CreatePaymentDetailsFail(
-                        fromPaymentProvider
-                      )
-                    );
-                  }
-                })
-              );
-          })
+          catchError(error =>
+            of(new fromActions.CreatePaymentDetailsFail(error))
+          )
         );
     })
   );
@@ -217,146 +169,11 @@ export class CheckoutEffects {
     })
   );
 
-  private domparser: DOMParser;
-
   constructor(
     private actions$: Actions,
     private cartDeliveryConnector: CartDeliveryConnector,
     private cartPaymentConnector: CartPaymentConnector,
     private occOrderService: OccOrderService,
     private productImageConverter: ProductImageNormalizer
-  ) {
-    if (typeof DOMParser !== 'undefined') {
-      this.domparser = new DOMParser();
-    }
-  }
-
-  private getPaymentSopResponseParams(
-    paymentDetails: any,
-    fromPaymentProvider: any,
-    mappingLabels: any
-  ) {
-    const sopResponseParams = {};
-
-    sopResponseParams['decision'] =
-      fromPaymentProvider[mappingLabels['hybris_sop_decision']];
-    sopResponseParams['amount'] =
-      fromPaymentProvider[mappingLabels['hybris_sop_amount']];
-    sopResponseParams['currency'] =
-      fromPaymentProvider[mappingLabels['hybris_sop_currency']];
-
-    sopResponseParams['billTo_country'] =
-      fromPaymentProvider[mappingLabels['hybris_billTo_country']];
-    sopResponseParams['billTo_firstName'] =
-      fromPaymentProvider[mappingLabels['hybris_billTo_firstname']];
-    sopResponseParams['billTo_lastName'] =
-      fromPaymentProvider[mappingLabels['hybris_billTo_lastname']];
-    sopResponseParams['billTo_street1'] =
-      fromPaymentProvider[mappingLabels['hybris_billTo_street1']];
-    sopResponseParams['billTo_city'] =
-      fromPaymentProvider[mappingLabels['hybris_billTo_city']];
-    sopResponseParams['billTo_postalCode'] =
-      fromPaymentProvider[mappingLabels['hybris_billTo_postalcode']];
-
-    sopResponseParams['card_cardType'] = paymentDetails.cardType.code;
-    sopResponseParams['card_accountNumber'] =
-      fromPaymentProvider[mappingLabels['hybris_sop_card_number']];
-    sopResponseParams['card_expirationMonth'] = paymentDetails.expiryMonth;
-    sopResponseParams['card_expirationYear'] = paymentDetails.expiryYear;
-    sopResponseParams['card_nameOnCard'] = paymentDetails.accountHolderName;
-    sopResponseParams['defaultPayment'] = paymentDetails.defaultPayment;
-    sopResponseParams['savePaymentInfo'] = true;
-
-    sopResponseParams['reasonCode'] =
-      fromPaymentProvider[mappingLabels['hybris_sop_reason_code']];
-    sopResponseParams['paySubscriptionCreateReply_subscriptionID'] =
-      fromPaymentProvider[mappingLabels['hybris_sop_subscriptionID']];
-
-    if (mappingLabels['hybris_sop_uses_public_signature'] === 'true') {
-      sopResponseParams[
-        'paySubscriptionCreateReply_subscriptionIDPublicSignature'
-      ] = fromPaymentProvider[mappingLabels['hybris_sop_public_signature']];
-    }
-
-    return sopResponseParams;
-  }
-
-  private getParamsForPaymentProvider(
-    paymentDetails: PaymentDetails,
-    parameters: { key; value }[],
-    mappingLabels: any
-  ) {
-    const params = this.convertToMap(parameters);
-    params[mappingLabels['hybris_account_holder_name']] =
-      paymentDetails.accountHolderName;
-    params[mappingLabels['hybris_card_type']] = paymentDetails.cardType.code;
-    params[mappingLabels['hybris_card_number']] = paymentDetails.cardNumber;
-    if (mappingLabels['hybris_combined_expiry_date'] === 'true') {
-      params[mappingLabels['hybris_card_expiry_date']] =
-        paymentDetails.expiryMonth +
-        mappingLabels['hybris_separator_expiry_date'] +
-        paymentDetails.expiryYear;
-    } else {
-      params[mappingLabels['hybris_card_expiration_month']] =
-        paymentDetails.expiryMonth;
-      params[mappingLabels['hybris_card_expiration_year']] =
-        paymentDetails.expiryYear;
-    }
-    params[mappingLabels['hybris_card_cvn']] = paymentDetails.cvn;
-
-    // billing address
-    params[mappingLabels['hybris_billTo_country']] =
-      paymentDetails.billingAddress.country.isocode;
-    params[mappingLabels['hybris_billTo_firstname']] =
-      paymentDetails.billingAddress.firstName;
-    params[mappingLabels['hybris_billTo_lastname']] =
-      paymentDetails.billingAddress.lastName;
-    params[mappingLabels['hybris_billTo_street1']] =
-      paymentDetails.billingAddress.line1 +
-      ' ' +
-      paymentDetails.billingAddress.line2;
-    params[mappingLabels['hybris_billTo_city']] =
-      paymentDetails.billingAddress.town;
-    params[mappingLabels['hybris_billTo_postalcode']] =
-      paymentDetails.billingAddress.postalCode;
-    return params;
-  }
-
-  private extractPaymentDetailsFromHtml(html: string): any {
-    const domdoc = this.domparser.parseFromString(html, 'text/xml');
-    const responseForm = domdoc.getElementsByTagName('form')[0];
-    const inputs = responseForm.getElementsByTagName('input');
-
-    const values = {};
-    for (let i = 0; inputs[i]; i++) {
-      const input = inputs[i];
-      if (
-        input.getAttribute('name') !== '{}' &&
-        input.getAttribute('value') !== ''
-      ) {
-        values[input.getAttribute('name')] = input.getAttribute('value');
-      }
-    }
-
-    // rejected for some reason
-    if (values['decision'] !== 'ACCEPT') {
-      const reason = { hasError: true };
-      Object.keys(values).forEach(name => {
-        if (name === 'reasonCode' || name.startsWith('InvalidField')) {
-          reason[name] = values[name];
-        }
-      });
-      return reason;
-    }
-
-    return values;
-  }
-
-  private convertToMap(paramList: { key; value }[]) {
-    return paramList.reduce(function(result, item) {
-      const key = item.key;
-      result[key] = item.value;
-      return result;
-    }, {});
-  }
+  ) {}
 }
